Move boxing logic into separate class
